Rename analysis as plugin
